Support become, become_method, and executable on the server side
Instead of dealing with parsing the command to identify when the other
side is attempting to use sudo, implement the logic for privilege
escalation on the server side and pass the various parameters on the
client.

The connection plugin now passes become, become_method, and executable
as additional parameters when they are present and these are used in
place of the defaults.