Implement ldap authentication with username and password over HTTP basic auth
Removing `remote_user` from the ansible.cfg file so it uses the current
user of the system for authentication. Adding `ask_pass` so it asks for
the password on every run. These will be needed when converting to this
and serve as an example of how to use this agent.