Better logging for authentication, support for SSL/TLS, and config file format
The authenticator now logs whatever it decides and also logs the name of
the person who authenticates on each request. This way, we can audit
what actions occur more easily.

We also support SSL/TLS for LDAP authentication so that can be secure.
If the host starts with `ldaps://`, it will automatically enable SSL.
You can also specify `ssl = true` and only put the host name as `host =
???`.

There is now an `[ldap]` section of the toml file to enable ldap. It is
disabled by default. All of the previous options which were hard-coded
are configurable from this section.