Exclude /ping from requiring authentication
Right now you can't exclude a handler from a specific route in martini.
We want only that endpoint to be open to everywhere since it's a health
check, but we don't want to take away the global auth handler. Even if
we wanted to explicitly mark each method that needed authentication, we
would have to restructure how authentication is installed on the server.

For now, just do a path check in the ldap auth handler and have it allow
the ping endpoint as a special exception. If we need to do this again,
it should be fixed permanently rather than using this same hack.