Add support for verifying client certificates
This allows using SSL verification for the clients themselves. This way,
you can verify what clients can connect to the agent instead of
potentially giving a hacker access to run commands as a privileged user.

This acts equivalent to the same encryption used by an SSH connection.
With the addition of password authentication with LDAP, this gives two
different methods of client verification.