Transfer the entirety of stdin for an exec to a file before pipelining
It's a security hole to execute a file over stdin through a daemon as if
the daemon spontaneously dies in the middle of transmission, the
interpreter will still partially execute the file.

Most Ansible modules try to prevent this by wrapping things in functions
and executing main at the bottom. This is a good solution and should
work in most cases, but we might as well execute the file from a file
rather than stdin if we're physically able to do that, which we are with
the daemon.